Fix on latest nightly, move from std::simd -> packed_simd
