+++ src/arch/unknown/intrin/abs.rs
@@ -7,7 +7,7 @@
 
 use crate::intrin::abs::Abs;
 use crate::arch::current::vecs::*;
-use crate::std::mem::transmute;
+use crate::core::mem::transmute;
 
 impl Abs for f32x4 {
     type Out = f32x4;
