+++ src/arch/x86/intrin/hsub.rs
@@ -11,7 +11,7 @@ use crate::arch::current::vecs::*;
 use crate::vecs::*;
 use crate::intrin::transmute::*;
 use crate::intrin::hsub::*;
-use crate::std::ops::Sub;
+use crate::core::ops::Sub;
 
 #[cfg(target_feature = "sse3")]
 impl HSub for f32x4 {
