+++ src/arch/x86/intrin/sum.rs
@@ -2,7 +2,7 @@ use crate::arch::current::vecs::*;
 use crate::vecs::*;
 use crate::vektor::x86_64::*;
 use crate::vektor::x86::*;
-use crate::std::ops::Add;
+use crate::core::ops::Add;
 use crate::intrin::upcast::Upcast;
 use crate::intrin::cmp::Cmp;
 use crate::intrin::abs::Abs;
