+++ src/intrin/macros.rs
@@ -37,7 +37,7 @@ macro_rules! rust_fallback_impl_binary {
                 #[inline(always)]
                 #[cfg(target_feature = $feat)]
                 fn $rustfn(&self, other: Self) -> Self {
-                    use crate::std::mem::transmute;
+                    use crate::core::mem::transmute;
                     optimized!();
                     unsafe { transmute($mmfn(transmute(*self), transmute(other), $($mmfnargs),*)) }
                 }
